fix(web/applications): pr tweaks (BOAS-508)
