+++ apps/web/src/server/applications/applications.types.d.ts
@@ -94,11 +94,3 @@ export interface DocumentationFile {
 	fileName: string;
 	url: string;
 }
-
-export interface DocumentationPageProps {
-	caseId: number;
-	folderId: number;
-	folderName: string | undefined;
-	folderTree: DocumentationCategory[] | null;
-	documentationFiles: DocumentationFile[] | null;
-}
