+++ apps/web/src/server/applications/pages/case/documentation/applications-documentation.types.d.ts
@@ -1,0 +1,9 @@
+import { DocumentationCategory, DocumentationFile } from '../../../applications.types';
+
+export interface DocumentationPageProps {
+	caseId: number;
+	folderId: number;
+	folderName: string | undefined;
+	folderTree: DocumentationCategory[] | null;
+	documentationFiles: DocumentationFile[] | null;
+}
