+++ apps/web/src/server/views/applications/case-documentation/project-documentation-folder.njk
@@ -9,8 +9,8 @@
 {% block sectionContent %}
 	{% set items = [] %}
 	{% set items = items | concat({ href: 'case-view'|url({caseId: caseId, step: 'project-documentation'}), text: 'Project documentation' }) %}
-	{% for aFolder in folderTree %}
-		{% set items = items | concat({ href: 'document-category'|url({caseId: caseId, documentationCategory: aFolder}), text: aFolder.displayNameEn }) %}
+	{% for folder in folderTree %}
+		{% set items = items | concat({ href: 'document-category'|url({caseId: caseId, documentationCategory: folder}), text: folder.displayNameEn }) %}
 	{% endfor %}
 
 	{{ govukBreadcrumbs({
